+++ assets/lodash/preact/src/examples/example-lodash.tsx
@@ -13,18 +13,11 @@ export class ExampleLodash extends Component<any, { result1: string }>  {
      */
     constructor(props?: any, context?: any) {
         super(props, context);
-        this.state = { result1: undefined };
-    }
 
-    /**
-     * The component is about to be mounted.
-     * @returns {void}
-     */
-    public componentWillMount(): void {
         const objA = { name: "bob", car: "porsche" };
         const objB = { name: "bill", age: 40 };
 
-        this.setState({ result1: JSON.stringify(_.assign(objA, objB)) });
+        this.state = { result1: JSON.stringify(_.assign(objA, objB)) };
     }
 
     /**